legacy-code/backend/backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.7
django-cors-headers>=4.3.1
channels>=4.0.0
daphne>=4.0.0
drf-nested-routers>=0.93.4
PyJWT>=2.8.0
cryptography>=41.0.5
Pillow>=10.0.0
python-dotenv>=1.0.0
